git-svn-id: svn://cybergarage.org:3960/project/uhttp/trunk@51 7ed65ffe-41cb-4a7d-888d-b4931f95f61e

+++ include/uhttp/http/HTTPSocket.h
@@ -79,12 +79,12 @@ public:
 
 private:
 
-	bool post(HTTPResponse *httpRes, const std::string &content, int contentOffset, int contentLength, bool isOnlyHeader, bool isChunked);
-	bool post(HTTPResponse *httpRes, uHTTP::InputStream *in, long contentOffset, long contentLength, bool isOnlyHeader, bool isChunked);
+	bool post(HTTPResponse *httpRes, const std::string &content, size_t contentOffset, size_t contentLength, bool isOnlyHeader, bool isChunked);
+	bool post(HTTPResponse *httpRes, uHTTP::InputStream *in, size_t contentOffset, size_t contentLength, bool isOnlyHeader, bool isChunked);
 
 public:
 
-	bool post(HTTPResponse *httpRes, int contentOffset, int contentLength, bool isOnlyHeader, bool isChunked);
+	bool post(HTTPResponse *httpRes, size_t contentOffset, size_t contentLength, bool isOnlyHeader, bool isChunked);
 	bool post(HTTPResponse *httpRes);
 	bool post(const std::string &content);
 	bool post(const char c);
